[GB.SDL] * BUG: Compute the framerate differently and more accurately. * BUG: Framerate limitation is more accurate now. * NEW: Window.Framerate property is a Float number now. * NEW: Manage the WAIT instruction. * NEW: The event loop can ignore input events, as requested by the WAIT   instruction, when it is called with no argument.
git-svn-id: svn://localhost/gambas/trunk@3772 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.sdl/src/Cwindow.h
@@ -60,12 +60,12 @@ typedef
 		bool openGL;
 		bool tracking;
 		// framerate control
-		Uint32 FPSLimit; // duration of a frame, if 0 -> no framerate limit
-		Uint32 lastTime;
+		double FPSLimit; // duration of a frame in milliseconds, if 0 -> no framerate limit
+		double lastTime;
 		// framerate count
 		Uint32 startTime;
 		Uint32 countFrames;
-		Uint32 currentFPS;
+		double currentFPS;
 	}
 	CWINDOW;
